refactor to move license validation into a seperate class
